+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+-   Show download progress now on top of game tile
+
 ## [0.5.2] - 2022-03-03
 
 ### Fixed