+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.0] - 2022-06-07
+
 ### Added
 
 -   Installer checksum verification
@@ -16,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 -   Show download progress now on top of game tile
+-   Updated dependencies
 
 ### Fixed
 
@@ -87,3 +90,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2018-07-16
 
 initial, very basic version with console interface
+
+[Unreleased]: https://github.com/NicoVIII/Basealize/compare/v0.6.0...HEAD
+[0.6.0]: https://github.com/NicoVIII/Basealize/compare/v0.5.2..v0.6.0
+[0.5.2]: https://github.com/NicoVIII/Basealize/compare/v0.5.1..v0.5.2
+[0.5.1]: https://github.com/NicoVIII/Basealize/compare/v0.5.0..v0.5.1
+[0.5.0]: https://github.com/NicoVIII/Basealize/compare/v0.4.0..v0.5.0
+[0.4.0]: https://github.com/NicoVIII/Basealize/compare/v0.3.0..v0.4.0
+[0.3.0]: https://github.com/NicoVIII/Basealize/compare/v0.2.0..v0.3.0
+[0.2.0]: https://github.com/NicoVIII/Basealize/compare/v0.1.0..v0.2.0
+[0.1.0]: https://github.com/NicoVIII/Basealize/releases/v0.1.0